Add comprehensive logging and validation in drawing context calls.
